+++ packages/core/RNSentryCocoaTester/RNSentryCocoaTesterTests/RNSentryTests.m
@@ -4,6 +4,7 @@
 #import <Sentry/SentryDebugImageProvider+HybridSDKs.h>
 #import <UIKit/UIKit.h>
 #import <XCTest/XCTest.h>
+@import Sentry;
 
 @interface RNSentryInitNativeSdkTests : XCTestCase
 
@@ -464,8 +465,13 @@ sucessfulSymbolicate(const void *, Dl_info *info)
     SentryOptions *sentryOptions = [[SentryOptions alloc] init];
     sentryOptions.debug = debug; // no local symbolication
 
+#if CROSS_PLATFORM_TEST
+    id sentrySDKMock = OCMClassMock([SentrySDKInternal class]);
+    OCMStub([(Class)sentrySDKMock options]).andReturn(sentryOptions);
+#else
     id sentrySDKMock = OCMClassMock([SentrySDK class]);
-    OCMStub([(SentrySDK *)sentrySDKMock options]).andReturn(sentryOptions);
+    OCMStub([(Class)sentrySDKMock options]).andReturn(sentryOptions);
+#endif
 
     id sentryDependencyContainerMock = OCMClassMock([SentryDependencyContainer class]);
     OCMStub(ClassMethod([sentryDependencyContainerMock sharedInstance]))