+++ packages/core/ios/RNSentry.mm
@@ -21,10 +21,22 @@
 #import <Sentry/PrivateSentrySDKOnly.h>
 #import <Sentry/SentryAppStartMeasurement.h>
 #import <Sentry/SentryBinaryImageCache.h>
+#import <Sentry/SentryBreadcrumb.h>
 #import <Sentry/SentryDebugImageProvider+HybridSDKs.h>
+#import <Sentry/SentryDebugMeta.h>
 #import <Sentry/SentryDependencyContainer.h>
+#import <Sentry/SentryEvent.h>
+#import <Sentry/SentryException.h>
 #import <Sentry/SentryFormatter.h>
-#import <Sentry/SentryOptions+HybridSDKs.h>
+#import <Sentry/SentryOptions.h>
+#import <Sentry/SentryUser.h>
+#if __has_include(<Sentry/SentryOptions+HybridSDKs.h>)
+#    define USE_SENTRY_OPTIONS 1
+#    import <Sentry/SentryOptions+HybridSDKs.h>
+#else
+#    define USE_SENTRY_OPTIONS 0
+#    import <Sentry/SentryOptionsInternal.h>
+#endif
 #import <Sentry/SentryScreenFrames.h>
 
 // This guard prevents importing Hermes in JSC apps
@@ -51,15 +63,7 @@
 
 #import "RNSentryExperimentalOptions.h"
 #import "RNSentryVersion.h"
-
-@interface
-SentrySDK (RNSentry)
-
-+ (void)captureEnvelope:(SentryEnvelope *)envelope;
-
-+ (void)storeEnvelope:(SentryEnvelope *)envelope;
-
-@end
+#import "SentrySDKWrapper.h"
 
 static bool hasFetchedAppStart;
 
@@ -107,7 +111,7 @@ RCT_EXPORT_METHOD(initNativeSdk
     [PrivateSentrySDKOnly addSdkPackage:REACT_NATIVE_SDK_PACKAGE_NAME
                                 version:REACT_NATIVE_SDK_PACKAGE_VERSION];
 
-    [SentrySDK startWithOptions:sentryOptions];
+    [SentrySDKWrapper startWithOptions:sentryOptions];
 
 #if TARGET_OS_IPHONE || TARGET_OS_MACCATALYST
     BOOL appIsActive =
@@ -243,8 +247,13 @@ RCT_EXPORT_METHOD(initNativeSdk
     [RNSentryReplay updateOptions:mutableOptions];
 #endif
 
+#if USE_SENTRY_OPTIONS
     SentryOptions *sentryOptions = [[SentryOptions alloc] initWithDict:mutableOptions
                                                       didFailWithError:errorPointer];
+#else
+    SentryOptions *sentryOptions = [SentryOptionsInternal initWithDict:mutableOptions
+                                                      didFailWithError:errorPointer];
+#endif
     if (*errorPointer != nil) {
         return nil;
     }
@@ -432,7 +441,11 @@ RCT_EXPORT_SYNCHRONOUS_TYPED_METHOD(NSString *, fetchNativePackageName)
 - (NSDictionary *)fetchNativeStackFramesBy:(NSArray<NSNumber *> *)instructionsAddr
                                symbolicate:(SymbolicateCallbackType)symbolicate
 {
+#if CROSS_PLATFORM_TEST
+    BOOL shouldSymbolicateLocally = [SentrySDKInternal.options debug];
+#else
     BOOL shouldSymbolicateLocally = [SentrySDK.options debug];
+#endif
     NSString *appPackageName = [[NSBundle mainBundle] executablePath];
 
     NSMutableSet<NSString *> *_Nonnull imagesAddrToRetrieveDebugMetaImages =
@@ -520,7 +533,7 @@ RCT_EXPORT_METHOD(fetchNativeDeviceContexts
     __block NSMutableDictionary<NSString *, id> *serializedScope;
     // Temp work around until sorted out this API in sentry-cocoa.
     // TODO: If the callback isnt' executed the promise wouldn't be resolved.
-    [SentrySDK configureScope:^(SentryScope *_Nonnull scope) {
+    [SentrySDKWrapper configureScope:^(SentryScope *_Nonnull scope) {
         serializedScope = [[scope serialize] mutableCopy];
 
         NSDictionary<NSString *, id> *user = [serializedScope valueForKey:@"user"];
@@ -724,7 +737,7 @@ RCT_EXPORT_METHOD(fetchViewHierarchy
 
 RCT_EXPORT_METHOD(setUser : (NSDictionary *)userKeys otherUserKeys : (NSDictionary *)userDataKeys)
 {
-    [SentrySDK configureScope:^(SentryScope *_Nonnull scope) {
+    [SentrySDKWrapper configureScope:^(SentryScope *_Nonnull scope) {
         [scope setUser:[RNSentry userFrom:userKeys otherUserKeys:userDataKeys]];
     }];
 }
@@ -773,7 +786,7 @@ RCT_EXPORT_METHOD(setUser : (NSDictionary *)userKeys otherUserKeys : (NSDictiona
 
 RCT_EXPORT_METHOD(addBreadcrumb : (NSDictionary *)breadcrumb)
 {
-    [SentrySDK configureScope:^(SentryScope *_Nonnull scope) {
+    [SentrySDKWrapper configureScope:^(SentryScope *_Nonnull scope) {
         [scope addBreadcrumb:[RNSentryBreadcrumb from:breadcrumb]];
     }];
 
@@ -787,12 +800,12 @@ RCT_EXPORT_METHOD(addBreadcrumb : (NSDictionary *)breadcrumb)
 
 RCT_EXPORT_METHOD(clearBreadcrumbs)
 {
-    [SentrySDK configureScope:^(SentryScope *_Nonnull scope) { [scope clearBreadcrumbs]; }];
+    [SentrySDKWrapper configureScope:^(SentryScope *_Nonnull scope) { [scope clearBreadcrumbs]; }];
 }
 
 RCT_EXPORT_METHOD(setExtra : (NSString *)key extra : (NSString *)extra)
 {
-    [SentrySDK
+    [SentrySDKWrapper
         configureScope:^(SentryScope *_Nonnull scope) { [scope setExtraValue:extra forKey:key]; }];
 }
 
@@ -802,7 +815,7 @@ RCT_EXPORT_METHOD(setContext : (NSString *)key context : (NSDictionary *)context
         return;
     }
 
-    [SentrySDK configureScope:^(SentryScope *_Nonnull scope) {
+    [SentrySDKWrapper configureScope:^(SentryScope *_Nonnull scope) {
         if (context == nil) {
             [scope removeContextForKey:key];
         } else {
@@ -813,17 +826,17 @@ RCT_EXPORT_METHOD(setContext : (NSString *)key context : (NSDictionary *)context
 
 RCT_EXPORT_METHOD(setTag : (NSString *)key value : (NSString *)value)
 {
-    [SentrySDK
+    [SentrySDKWrapper
         configureScope:^(SentryScope *_Nonnull scope) { [scope setTagValue:value forKey:key]; }];
 }
 
-RCT_EXPORT_METHOD(crash) { [SentrySDK crash]; }
+RCT_EXPORT_METHOD(crash) { [SentrySDKWrapper crash]; }
 
 RCT_EXPORT_METHOD(closeNativeSdk
                   : (RCTPromiseResolveBlock)resolve rejecter
                   : (RCTPromiseRejectBlock)reject)
 {
-    [SentrySDK close];
+    [SentrySDKWrapper close];
     resolve(@YES);
 }
 
@@ -1026,7 +1039,7 @@ RCT_EXPORT_METHOD(crashedLastRun
                   : (RCTPromiseResolveBlock)resolve rejecter
                   : (RCTPromiseRejectBlock)reject)
 {
-    resolve(@([SentrySDK crashedLastRun]));
+    resolve(@([SentrySDKWrapper crashedLastRun]));
 }
 
 // Thanks to this guard, we won't compile this code when we build for the old architecture.