+++ samples/react-native/src/App.tsx
@@ -177,6 +177,7 @@ Sentry.init({
   replaysSessionSampleRate: 1.0,
   replaysOnErrorSampleRate: 1.0,
   ignoreErrors: ['should', /(.)*2(.)*/],
+  replaysSessionQuality: 'medium', // default
   spotlight: true,
   // This should be disabled when manually initializing the native SDK
   // Note that options from JS are not passed to the native SDKs when initialized manually